added Scraper interface
